# shop-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/credit-cards": {
            "get": {
                "tags": [
                    "CreditCard"
                ],
                "summary": "getCreditCardList",
                "description": "Get all CreditCards",
                "operationId": "237c0bfdf6092acf25f08761cfbe1602",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CreditCard"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CreditCard"
                ],
                "summary": "createCreditCard",
                "description": "Create CreditCard",
                "operationId": "b2a0f322a397b706a231709840da7f2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreditCard"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CreditCard"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/credit-cards/{id}": {
            "get": {
                "tags": [
                    "CreditCard"
                ],
                "summary": "getCreditCardItem",
                "description": "Get CreditCard",
                "operationId": "c0601dd20ce0e2f7748d285544517c72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CreditCard",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CreditCard"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CreditCard"
                ],
                "summary": "updateCreditCard",
                "description": "Update CreditCard",
                "operationId": "90dab335014a0dde7daaf8f78ebf672d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CreditCard",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreditCard"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CreditCard"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CreditCard"
                ],
                "summary": "deleteCreditCard",
                "description": "Delete CreditCard",
                "operationId": "9a695a3467161c781d331791a3789662",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CreditCard",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreditCard": {
                "required": [
                    "user_id",
                    "card_number",
                    "cardholder_name",
                    "expiration_date",
                    "cvv",
                    "balance"
                ],
                "properties": {
                    "card_number": {
                        "description": "16-digit card number",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "cardholder_name": {
                        "description": "Name on the card",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "expiration_date": {
                        "description": "Expiration date of card",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": false
                    },
                    "cvv": {
                        "description": "3-digit CVV",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "balance": {
                        "description": "Available balance",
                        "type": "number",
                        "format": "number",
                        "readOnly": false,
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}